term_track/pyproject.toml
[tool.poetry]
name = "term-track"
version = "0.1.0"
description = ""
authors = ["tanupat <tanupat92@gmail.com>"]
readme = "README.md"
packages = [{include = "term_track"}]

[tool.poetry.dependencies]
pymongo = "^4.6.0"
pandas = "^2.1.3"
tqdm = "^4.66.1"
python = ">=3.10"
streamlit = "^1.28.2"
python-dotenv = "^1.0.0"
streamlit-ext = "^0.1.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
